hbase deltetable flag added

+++ CryptoAnalysis/src/main/java/Main.java
@@ -41,56 +41,55 @@ import scala.Tuple2;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 public class Main {
-	
-	private static final String TABLE_NAME = "cryptomarketdata";
-	
-	// to check if table has been created and initialized so only update happens after first RDD create a table
-	private static int tableInitialized = 0;
-	
-	public static void putDataIntoHBase (String cryptoName, double averagePrice) throws IOException {
-    	Configuration hbaseConfig = HBaseConfiguration.create();
-		try (Connection connection = ConnectionFactory.createConnection(hbaseConfig);
-		Admin admin = connection.getAdmin())
-		{
-			TableName tableName = TableName.valueOf(TABLE_NAME);
-			
-			if(tableInitialized == 0) {
-				admin.disableTable(tableName);
-				admin.deleteTable(tableName);
-			}
-			
-			if (!admin.tableExists(tableName))
-			{
-				tableInitialized++;
-				HTableDescriptor table = new HTableDescriptor(tableName);
-				table.addFamily(new HColumnDescriptor("cf"));
-				
-				System.out.println("Creating table....");
-				admin.createTable(table);
-				System.out.println("Done");
-			} else if(tableInitialized == 0) {
-				System.out.println("Table Exist.....");
-			}
 
-			
-			Table t = connection.getTable(tableName);
-			long timestamp = System.currentTimeMillis();
-			Put put = new Put(Bytes.toBytes(cryptoName));
+    private static final String TABLE_NAME = "cryptomarketdata";
+    private static final boolean deleteTable = false;
 
-			// set average price column with timestamp to store price history
-			put.addColumn(Bytes.toBytes("cf"), Bytes.toBytes("AveragePrice:"+timestamp), Bytes.toBytes(String.valueOf(averagePrice)));
-			t.put(put);
-		}
-	}
+    // to check if table has been created and initialized so only update happens
+    // after first RDD create a table
+    private static int tableInitialized = 0;
+
+    public static void putDataIntoHBase(String cryptoName, double averagePrice) throws IOException {
+        Configuration hbaseConfig = HBaseConfiguration.create();
+        try (Connection connection = ConnectionFactory.createConnection(hbaseConfig);
+                Admin admin = connection.getAdmin()) {
+            TableName tableName = TableName.valueOf(TABLE_NAME);
+
+            if (tableInitialized == 0 && deleteTable) {
+                admin.disableTable(tableName);
+                admin.deleteTable(tableName);
+            }
+
+            if (!admin.tableExists(tableName)) {
+                tableInitialized++;
+                HTableDescriptor table = new HTableDescriptor(tableName);
+                table.addFamily(new HColumnDescriptor("cf"));
+
+                System.out.println("Creating table....");
+                admin.createTable(table);
+                System.out.println("Done");
+            } else if (tableInitialized == 0) {
+                System.out.println("Table Exist.....");
+            }
+
+            Table t = connection.getTable(tableName);
+            long timestamp = System.currentTimeMillis();
+            Put put = new Put(Bytes.toBytes(cryptoName));
+
+            // set average price column with timestamp to store price history
+            put.addColumn(Bytes.toBytes("cf"), Bytes.toBytes("AveragePrice:" + timestamp),
+                    Bytes.toBytes(String.valueOf(averagePrice)));
+            t.put(put);
+        }
+    }
 
     public static void main(String[] args) throws IOException {
         // Set up Spark Streaming configuration
         SparkConf sparkConf = new SparkConf().setAppName("KafkaSparkStreamingConsumer").setMaster("local[*]");
         JavaStreamingContext streamingContext = new JavaStreamingContext(sparkConf, Durations.seconds(10));
-        
+
         // checkpoint for window streaming
         streamingContext.checkpoint("hdfs://localhost:8020/cloudera/cloudera/spark_checkpoint/");
-        
 
         // Kafka configuration
         String kafkaBrokers = "localhost:9092";
@@ -117,22 +116,24 @@ public class Main {
             ObjectMapper objectMapper = new ObjectMapper();
             return objectMapper.readValue(jsonData, CryptoDataResponse.class);
         }).flatMap(d -> d.getData().iterator());
-        
-        // Create a pair DStream with cryptocurrency name as key and a tuple of price and count as value
-        JavaPairDStream<String, Tuple2<Double, Integer>> pairDStream = coinMarketDataStream.mapToPair(data -> new Tuple2<>(data.getSymbol(), new Tuple2<>(data.getQuote().getUsd().getPrice(), 1)));
+
+        // Create a pair DStream with cryptocurrency name as key and a tuple of price
+        // and count as value
+        JavaPairDStream<String, Tuple2<Double, Integer>> pairDStream = coinMarketDataStream.mapToPair(
+                data -> new Tuple2<>(data.getSymbol(), new Tuple2<>(data.getQuote().getUsd().getPrice(), 1)));
 
         // Use the reduceByKeyAndWindow transformation to accumulate values and count
         JavaPairDStream<String, Tuple2<Double, Integer>> reducedStream = pairDStream.reduceByKeyAndWindow(
-            (v1, v2) -> new Tuple2<>(v1._1() + v2._1(), v1._2() + v2._2()),
-            (v1, v2) -> new Tuple2<>(v1._1() - v2._1(), v1._2() - v2._2()),
-            Durations.seconds(60),
-            Durations.seconds(60)
-        );
+                (v1, v2) -> new Tuple2<>(v1._1() + v2._1(), v1._2() + v2._2()),
+                (v1, v2) -> new Tuple2<>(v1._1() - v2._1(), v1._2() - v2._2()),
+                Durations.seconds(60),
+                Durations.seconds(60));
 
         // Calculate the average for each cryptocurrency
         JavaPairDStream<String, Double> averageStream = reducedStream.mapValues(value -> value._1() / value._2());
 
-        // store the average price for each cryptocurrency into hbase with current timestamp
+        // store the average price for each cryptocurrency into hbase with current
+        // timestamp
         averageStream.foreachRDD(rdd -> {
             rdd.foreach(tuple -> {
                 String cryptoName = tuple._1();
@@ -140,8 +141,6 @@ public class Main {
                 putDataIntoHBase(cryptoName, averagePrice);
             });
         });
-
-  
 
         // Start the Spark Streaming context
         streamingContext.start();
@@ -153,6 +152,5 @@ public class Main {
             e.printStackTrace();
         }
     }
-    
 
 }
